+++ database/factories/ActionInstanceFieldFactory.php
@@ -20,7 +20,7 @@ use Faker\Generator as Faker;
 
 $factory->define(ActionInstanceField::class, function(Faker $faker) {
     return [
-        'event_field' => $faker->word,
+        'action_value' => $faker->word,
         'action_field' => $faker->text,
         'action_instance_id' => function() {
             return factory(ActionInstance::class)->create()->id;