+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0] - (18/03/2020)
+
+### Changed
+- Action Instances, Activities, Module Instances and Logics retrieve the user ID from UserAuthentication not Authentication
+- Changed event_field in action_instance_fields table to action_value
+- Allow text to be entered into the action value, replace any {{event:field}} with the event 'field' value
+- Added static Action::options() method to return a form schema
+- ActionInstance::action_schema property
+
+### Removed
+- getFields and getFieldMetaData from Action contract
+- ActionInstance::action_fields property
+
 ## [3.1.2] - (18/03/2020)
 
 ### Fixed
@@ -127,7 +140,8 @@ BristolSU\Support\Testing\HandlesAuthorization to help with testing modules with
 ### Added
 - Initial Release
 
-[Unreleased]: https://github.com/bristol-su/support/compare/v3.1.1...HEAD
+[Unreleased]: https://github.com/bristol-su/support/compare/v4.0...HEAD
+[4.0]: https://github.com/bristol-su/support/compare/v3.1.1...v4.0
 [3.1.1]: https://github.com/bristol-su/support/compare/v3.1...v3.1.1
 [3.1]: https://github.com/bristol-su/support/compare/v3.0.2...v3.1
 [3.0.2]: https://github.com/bristol-su/support/compare/v3.0.1...v3.0.2